refactor user

+++ src/main/java/com/fish/myspringboot/mapper/UserMapper.java
@@ -4,12 +4,13 @@ import com.fish.myspringboot.pojo.User;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Mapper
 public interface UserMapper {
 
-    User getById(@Param("id") int id);
+    User getById(@Param("id") BigInteger id);
 
     int updateUser(User user);
